requirements.txt
numpy
matplotlib
yaml
pillow
tqdm
scipy
lmdb
torch
torchvision
six